Switched DatumAccess to use a sane key.
Previously it was possible (and indeed happened) that multiple DatumAccess rows
were created for a single datum. My aversion to weird compound string keys
motivated me to do something even more crack-smoking which was to not have the
primary key of the DatumAccess row be the cortexId and datumId to which it
applied.

That situation is now remedied (and this commit masks the series of
intermediate changes I made in order to migrated the existing DatumAccess rows
to the new DatumAccess rows).