A bunch of yak shaving.
GAE demanded that I upgrade to Java 8. Simple enough, just rebuild with a line
added to appengine-web.xml. Oh wait, rebuild.

Objectify used to be hosted on Google Code with its own custom Maven repo, yay.
That of course has disappeared some time in the last five years and my cache of
it also disappeared in the last couple of years. Alas.

So we upgrade to Objectify 5.x which of course involves major API changes and
flip flopping on a bunch of highly important semantic defaults. Yay.

While I'm in here mucking about, I updated the GAE SDK (pretty painless, thanks
Google!) and Scala (also pretty painless, thanks Lightbend!).

I opted *not* to upgrade to GWT 2.8.1 because I'm 99% sure that would not be
painless, and so I'm going to leave that pain for some future yak to demand.